overwhelmingly card generating

+++ UnoSharp/GameComponent/Card.cs
@@ -76,9 +76,9 @@ namespace UnoSharp
             }
         }
 
-        public static IEnumerable<Card> Generate(int num)
+        public static IEnumerable<Card> Generate(ulong num)
         {
-            for (var i = 0; i < num; i++)
+            for (var i = 0ul; i < num; i++)
             {
                 yield return Generate();
             }
